Merge in CONNECT/wisetime-connector-java from feature/WT-9831-clio-connector-re-authenticate-message to master

* commit 'e4bffbdae1092388855d4adab42888951497a35c':
  WT-9831 add javadoc
  WT-9831 add health check failure endpoints

+++ src/test/java/io/wisetime/connector/api_client/DefaultApiClientIntegrationTest.java
@@ -199,7 +199,7 @@ class DefaultApiClientIntegrationTest {
     managedConfigRequest.setEnvironment(RuntimeEnvironmentUtil.getEnvProperties());
     managedConfigRequest.connectorType("test_connector_type");
     managedConfigRequest.setConnectorLibraryVersion(RuntimeEnvironmentUtil.getLibraryImplVersion());
-    managedConfigRequest.clientTimestamp((double) Instant.now().getEpochSecond());
+    managedConfigRequest.clientTimestamp(Instant.now().getEpochSecond());
     ManagedConfigResponse configResponse = defaultApiClient.getTeamManagedConfig(managedConfigRequest);
     log.info(configResponse.toString());
   }
